fix(compilation): Complete remaining compilation fixes from multi-agent coordination
- Fix final LLM optimized controller and pipeline issues
- Complete monitoring alerting router implementations
- Finalize Nephio workload and workflow engine fixes
- Complete performance analyzer and benchmark fixes
- Finish security CA remaining interface implementations
- Complete validation test framework final fixes

+++ pkg/performance/goroutine_pools.go
@@ -798,7 +798,7 @@ func (pool *EnhancedGoroutinePool) GetMetrics() PoolMetrics {
 		Throughput:         pool.metrics.Throughput,
 		CPUUtilization:     pool.metrics.CPUUtilization,
 		MemoryUsage:        atomic.LoadInt64(&pool.metrics.MemoryUsage),
-		GoroutineCount:     int(atomic.LoadInt64(&pool.metrics.GoroutineCount)),
+		GoroutineCount:     int64(atomic.LoadInt64(&pool.metrics.GoroutineCount)),
 		ScalingEvents:      atomic.LoadInt64(&pool.metrics.ScalingEvents),
 		PreemptionEvents:   atomic.LoadInt64(&pool.metrics.PreemptionEvents),
 		AffinityViolations: atomic.LoadInt64(&pool.metrics.AffinityViolations),
